+++ org.eclipse.lsp4j.debug/src/main/xtend-gen/org/eclipse/lsp4j/debug/RestartFrameArguments.java
@@ -25,21 +25,21 @@ public class RestartFrameArguments {
    * Restart this stackframe.
    */
   @NonNull
-  private Long frameId;
+  private Integer frameId;
   
   /**
    * Restart this stackframe.
    */
   @Pure
   @NonNull
-  public Long getFrameId() {
+  public Integer getFrameId() {
     return this.frameId;
   }
   
   /**
    * Restart this stackframe.
    */
-  public void setFrameId(@NonNull final Long frameId) {
+  public void setFrameId(@NonNull final Integer frameId) {
     this.frameId = Preconditions.checkNotNull(frameId, "frameId");
   }
